\name{BASIS}
\alias{BASIS}
\docType{data}
\title{An Example Data File for the Gauss Model}
\description{This is a 1000x481 sample feature matrix}
\usage{data(BASIS)}
\format{
  The format is:
 int [1:1000, 1:481] 0 -1 0 0 1 0 1 0 1 0 ...
}
\details{The data was simulated on a 2400cM chromosome, each column corresponded to an even spaced QTL}
\source{ 	Huang, A., Xu, S., and Cai, X. (2014). Empirical Bayesian elastic net for multiple quantitative trait locus mapping. Heredity 10.1038/hdy.2014.79 \cr
}

\examples{
data(BASIS)
}
\keyword{datasets}
