\name{EBEN-package}
\alias{EBEN-package}
\alias{EBEN}
\docType{package}
\title{Empirical Bayesian Elastic Net (EBEN)}

\description{ Fast EBEN  model implemented with memory efficient c code. \cr
			LAPACK/BLAS are used for most linear algebra computations. \cr
			Linear regression model.
	}
\details{
\tabular{ll}{
Package: \tab EBEN\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-28\cr
License: \tab gpl\cr
}

}
\author{
Anhui Huang; Shizhong Xu; Xiaodong Cai
Maintainer: Anhui Huang <a.huang1@umiami.edu>
}
\references{
Huang, A., Xu, S., and Cai, X. (2013). Empirical Bayesian elastic net for multiple quantitative trait locus mapping. submitted}
\keyword{ package }
