\name{y}
\alias{y}
\docType{data}
\title{Sample Response Data for Gaussian Model}
\description{Corresponding to the response of BASIS}

\usage{data(y)}
\format{
  The format is:
 num [1:1000, 1] 113.5 97.1 116.6 96.7 105.5 ...
}

\source{ Huang, A., Xu, S., and Cai, X. (2013). Empirical Bayesian elastic net for multiple quantitative trait locus mapping. submitted}

\examples{
data(y)
}
\keyword{datasets}
